+++ Examples/Resources/SPOTenantCDNPolicy/2-Update.ps1
@@ -14,7 +14,7 @@ Configuration Example
 
     node localhost
     {
-        SPOTenantCDNPolicy 'PublicCDNPolicy'
+        SPOTenantCdnPolicy 'PublicCdnPolicy'
         {
             IncludeFileExtensions                = @('.jpg', '.png')
             CDNType                              = "Public"